Support local webserver to install VIC UI plugin
In case the vCenter or VCSA has a local web server that it can access
user will provide this information in the configs so that
the installer script can either upload expanded plugin files to
the vCenter's plugin cache folder OR proceed with the standard plugin
deployment procedure where the plugin bundles are fetched from
print (if using HTTPS). For this matter, I have also updated the
build-deployable.xml build script under the ui/vic-ui/ so that upon
running the build script both the zip file and the expanded contents
will be copied to ui/installer/vsphere-client-serenity.