[WIP] Improve UI plugin installer for VCSA
We no longer require separate installer scripts for H5 Client plugin
and Flex Client plugin. Also in line with the OVA work I dropped the support for
NOURL (a.k.a no-webserver way to install plugin). This might eventually be
introduced back if we want to keep two options going in parallel (Bintray and
OVA)

NOTE: Don't forget to regenerate .drone.sec as .drone.yml was modified