Show datastores only belonging to the selected compute resource (#27)
* [WIP] Use datastores only attached to cluster

* Fix typo

* Show only datastores linked to the chosen cluster

* Remove unnecessary imports from wizard module

Also update the logic for updating the datastores select input
as the user changes the compute resource in the Compute Capacity
section.

* Set endpoint VM memory max value up to maxUsage

This commit prevents the next button from getting disabled
automatically in the Compute Capacity section when the user selects
a host or cluster that has less than 2048MB of ram available for
VM endpoint. As a solution, we set the smaller value between 2048
and memory['maxUsage'] so that the validation doesn't disable the
next button after selecting a host/cluster.